+++ Tests/ResponsibleTestBase.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using NUnit.Framework;
 using Responsible.Tests.Utilities;
@@ -47,7 +48,8 @@ namespace Responsible.Tests
 				this.Scheduler,
 				this.ExternalResultSource(),
 				this.FailureListener,
-				this.GlobalContextProvider);
+				this.GlobalContextProvider,
+				this.RethrowableExceptions);
 		}
 
 		[TearDown]
@@ -59,5 +61,6 @@ namespace Responsible.Tests
 		protected virtual IExternalResultSource ExternalResultSource() => null;
 		protected virtual IFailureListener MakeFailureListener() => null;
 		protected virtual IGlobalContextProvider MakeGlobalContextProvider() => null;
+		protected virtual IReadOnlyList<Type> RethrowableExceptions => null;
 	}
 }